requirements.txt
streamlit>=1.30.0
azure-ai-projects>=2.0.0
azure-identity>=1.16.0
msal>=1.28.0
python-dotenv>=1.0.0
